first implementation for BeatInfo service with deserialization and message / data classes, added new methods to convert long and double values from byte array in Utils.java added few lombok annotations

+++ Backend/src/main/java/com/openstreamingtools/backend/messages/stagelinqmessages/BeatData.java
@@ -9,5 +9,5 @@ public class BeatData {
     private double beat;
     private double totalBeats;
     private double BPM;
-    private int samples;
+    private double samples;
 }
